Moved data label functions of pie series into composition.
